Integrated phone number search logic with account search
